Use new SamplesTree structure in the code
